Extract bucket.rb to its own file.
